requirements.txt
fastapi==0.87.0
fastapi-pagination==0.11.4
uvicorn==0.19.0
pydantic==1.9.1
psycopg2==2.9.5
marshmallow-sqlalchemy==0.26.1